fix bug round value in video report

+++ src/Tagcade/Service/Report/VideoReport/Selector/Grouper/Groupers/WaterfallTagGrouper.php
@@ -150,7 +150,7 @@ class WaterfallTagGrouper extends AbstractGrouper
     protected function addEstSupplyCost($estSupplyCost)
     {
         if (!is_null($estSupplyCost)) {
-            $this->estSupplyCost += (int) $estSupplyCost;
+            $this->estSupplyCost += (float) $estSupplyCost;
         }
     }
